test: extract turn skill context suites
