+++ scripts/rust-source-size-baseline.txt
@@ -17,7 +17,7 @@ crates/agent-engine/src/runtime/prompt_cache.rs 2233
 crates/agent-engine/src/runtime/submission_handlers.rs 2981
 crates/agent-engine/src/runtime/tool_orchestrator.rs 3291
 crates/agent-engine/src/runtime/tool_policy.rs 1107
-crates/agent-engine/src/runtime/turn_executor.rs 2447
+crates/agent-engine/src/runtime/turn_executor.rs 1885
 crates/agent-engine/src/runtime/virtual_tools.rs 2476
 crates/agent-engine/src/runtime/virtual_tools_tests.rs 3331
 crates/agent-engine/src/skills/injector.rs 2490
